requirements.txt
Django==4.2.3
psycopg2>=2.8
djangorestframework==3.15.2
django-filter==2.4.0
djangorestframework-simplejwt==4.7.2
drf-spectacular==0.27.0